v1.3.1
Hotfix for the v1.3 release. 

I adjusted the cursor coordinates to be farther to the right to ensure the whole field is highlighted before the fade in or out values are sent. 

Along with adding sleep commands or the script will run too quickly in certain sections of the fade in or out hotkeys. Resulting in unintentional clicks on the cursor coordinates on the audio library tab instead.